+++ js/application.js
@@ -11,13 +11,19 @@ const ControllerId = {
 };
 
 const saveState = (state) => {
-  return JSON.stringify(state);
+  const stringified = JSON.stringify(state);
+  console.log(stringified);
+  return stringified;
 };
 
 const loadState = (dataString) => {
+console.log(`4: передано в распарс - ${dataString}`);
   try {
-    return JSON.parse(dataString);
+    const parsed = JSON.parse(dataString);
+    console.log(`5: передаются данные для формирования вью - ${parsed}`);
+    return parsed;
   } catch (e) {
+    console.log(e);
     return initialState;
   }
 };
@@ -33,6 +39,7 @@ export default class Application {
   static init() {
     const hashChangeHandler = () => {
       const hashValue = location.hash.replace(`#`, ``);
+      console.log(`1: записывает в хэш - ${hashValue}`);
       const [id, data] = hashValue.split(`?`);
       this.changeHash(id, data);
     };
@@ -42,6 +49,8 @@ export default class Application {
 
   static changeHash(id, data) {
     const controller = routes[id];
+    console.log(`2: ищем контролер - ${controller}`);
+    console.log(`3: дата которая будет передана в контролер - ${data}`);
     if (controller) {
       controller.init(loadState(data));
     }
