+++ js/game/game-model.js
@@ -1,13 +1,14 @@
 
 import {
   generateArtistQuestion, generateGenreQuestion, tick, reset, nextLevel, addMistake, addAnswer,
-  answersStack, resultScreenData
+  answersStack, resultScreenData, initialState
 } from "../data/data";
 
 export default class GameModel {
-  constructor(stack = answersStack, stats = resultScreenData) {
+  constructor(state = initialState, stack = answersStack, stats = resultScreenData.win) {
     this.stack = stack;
     this.stats = stats;
+    this.state = state;
   }
 
   update(newState) {
@@ -36,13 +37,13 @@ export default class GameModel {
   }
 
   getArtistLevel() {
-    this.data = generateArtistQuestion();
-    return this.data;
+    this.state.question = generateArtistQuestion();
+    return this.state.question;
   }
 
   getGenreLevel() {
-    this.data = generateGenreQuestion();
-    return this.data;
+    this.state.question = generateGenreQuestion();
+    return this.state.question;
   }
 
   get time() {
